Improve based on the latest Core and TCK PR
